# gen/openapiv2/headscale/v1/headscale.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "headscale/v1/headscale.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/machine/{machineId}": {
      "get": {
        "operationId": "HeadscaleService_GetMachine",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetMachineResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "machineId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "HeadscaleService"
        ]
      },
      "delete": {
        "operationId": "HeadscaleService_DeleteMachine",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMachineResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "machineId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "HeadscaleService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1DeleteMachineResponse": {
      "type": "object"
    },
...
    "v1GetMachineResponse": {
      "type": "object",
      "properties": {
        "machine": {
          "$ref": "#/definitions/v1Machine"
        }
      }
    },
...
    "v1Machine": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "machineKey": {
          "type": "string"
        },
        "nodeKey": {
          "type": "string"
        },
        "discoKey": {
          "type": "string"
        },
        "ipAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "namespace": {
          "$ref": "#/definitions/v1Namespace"
        },
        "lastSeen": {
          "type": "string",
          "format": "date-time"
        },
        "lastSuccessfulUpdate": {
          "type": "string",
          "format": "date-time"
        },
        "expiry": {
          "type": "string",
          "format": "date-time"
        },
        "preAuthKey": {
          "$ref": "#/definitions/v1PreAuthKey"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "registerMethod": {
          "$ref": "#/definitions/v1RegisterMethod"
        }
      }
    },
...
    "v1Namespace": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1PreAuthKey": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "reusable": {
          "type": "boolean"
        },
        "ephemeral": {
          "type": "boolean"
        },
        "used": {
          "type": "boolean"
        },
        "expiration": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1RegisterMethod": {
      "type": "string",
      "enum": [
        "REGISTER_METHOD_UNSPECIFIED",
        "REGISTER_METHOD_AUTH_KEY",
        "REGISTER_METHOD_CLI",
        "REGISTER_METHOD_OIDC"
      ],
      "default": "REGISTER_METHOD_UNSPECIFIED"
    },
...
  }
}